requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.2.0
cryptography==36.0.2
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
MarkupSafe==3.0.2
numpy==1.26.4
nvidia-nccl-cu12==2.26.5
pandas==2.2.3
pillow==11.2.1
pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
rdkit-pypi==2022.9.5
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
SQLAlchemy==2.0.40
threadpoolctl==3.6.0
typing_extensions==4.13.2
tzdata==2025.2
Werkzeug==3.1.3
xgboost==3.0.0
